ServidorSubSql/requirementsPUBSQL.txt
Flask
Flask-sqlalchemy
Flask-Migrate
gunicorn==19.9.0
PyMySQL==0.9.3
SQLAlchemy==1.2.17
google-cloud